backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
distlib==0.3.8
dnspython==2.6.1
fastapi==0.115.0
filelock==3.14.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
motor==3.5.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.8.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.30.6
virtualenv==20.26.2
openai==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5




